fix(build): count the modules the `imports` ratchet skips and keep a failed run's stage
- Count the candidates the baseline's uncomparable set kept out and show them as `skipped`
- Zero that count on a pointed `PROSE_IMPORTS_MODULE` run, which narrows for a different reason
- Keep the stage when a run fails on unmeasured modules, which discarded its tree
- Split the 762-line `tests.rs` into `tests/`, one file per surface with `broken` in `mod.rs`
- Drop the vestigial `interpreter` step from `🪟 Imports`, whose outputs no step there reads
- Pair `PYTHON` with `PYTHON_VAR` in `execute.rs` and rename `corpus::interpreter` to `standard_library`
- Resolve the interpreter inside `Runner::new`, dropping a parameter from it and `Sweep::new`
- Collapse `bake`'s two width maps onto a `keyed` helper taking one projection
- Drop the unused `Clone` on `Carried` and `Eq`/`PartialEq` on `Baseline`, and reach `BTreeSet::first`
- Pin the uncomparable bucket, a malformed set, the width knob, and one frame's two modules
- Correct `CONTRIBUTING.md`, which said the gate fails on a fresh break alone
