+++ .mise/tasks/gha/brief
@@ -123,7 +123,7 @@ class Summary:
 
     def _row(self, jobs: dict, job: str) -> dict:
         """
-        The mark and the name one corpus row shows.
+        Return the mark and the name one corpus row shows.
 
         Args:
             jobs : The workflow's job table.
