+++ CONTRIBUTING.md
@@ -137,11 +137,11 @@ Every corpus task defaults to the interpreter's own standard library and takes a
 
 `rust:mutations` mutates the corpus `rust:settle` sweeps and formats each variant with the unstable-output notice on, so a rewrite that settles over the corpus as written but not over a parseable edit of it still surfaces. Each mutation edits the module's concrete syntax tree, reordering top-level statements and class-body members, widening and narrowing identifiers, wrapping a sample of calls' arguments in redundant parentheses, injecting comments and suppression directives, and converting the line endings to CRLF, with every variant compiled before it lands so the report names a defect in *Prose* rather than one the mutation introduced. A second positional argument moves the pass off its sixty-second budget, and `PROSE_SETTLE_MUTATIONS` names the directory the set lives in, one `rust:bake` wrote or one the pass writes itself when the directory is absent. The pass formats that set in place, so a directory reused across runs carries the previous run's output rather than the mutated input, and the regeneration line the task prints rebuilds the input. A module the tree parser does not model, a template string among them, yields no variant, so read the denominator on every count before comparing two runs.
 
-`rust:imports` answers the question the settle probes cannot, whether a rewrite that settles still imports, by copying the corpus, formatting the copy, and executing every module the formatter rewrote from both trees in a fresh interpreter with its own tree first on `sys.path`, so a relative import resolves and a rewrite of one module reaches every module importing it. It runs as a `cargo test` target beside the settle and subset sweeps, formatting through the library rather than the binary, and the only Python it carries is the probe that loads one module inside the interpreter under test and reports the names it bound. A module counts as broken where the original imports cleanly and the formatted copy raises, times out, or binds a different namespace, with a second run of both sides confirming it, and the report keys each break by the frame it raises in, why it broke, and the rules whose fixes cover that line or dropped the binding of the name it misses, carrying one line per distinct defect with the modules it reached counted, the diff around the row it names, and the one-module command that reproduces it, leaving out every break the baseline already carries. `PROSE_IMPORTS_PYTHON` names the interpreter whose standard library the sweep runs, `PROSE_SETTLE_WIDTHS` adds widths beside the default, `PROSE_IMPORTS_TIMEOUT` bounds one module's run, and `PROSE_IMPORTS_BAKE` writes the break set into the file it names. Every other run ratchets against the set tracked at `crate/tests/imports/baseline.json`, failing only on a break that set does not carry and failing outright where nothing readable sits there.
+`rust:imports` answers the question the settle probes cannot, whether a rewrite that settles still imports, by copying the corpus, formatting the copy, and executing every module the formatter rewrote from both trees in a fresh interpreter with its own tree first on `sys.path`, so a relative import resolves and a rewrite of one module reaches every module importing it. It runs as a `cargo test` target beside the settle and subset sweeps, formatting through the library rather than the binary, and the only Python it carries is the probe that loads one module inside the interpreter under test and reports the names it bound. A module counts as broken where the original imports cleanly and the formatted copy raises, times out, or binds a different namespace, with a second run of both sides confirming it, and the report keys each break by the frame it raises in, why it broke, and the rules whose fixes cover that line or dropped the binding of the name it misses, carrying one line per distinct defect with the modules it reached counted, the diff around the row it names, and the one-module command that reproduces it, leaving out every break the baseline already carries. `PROSE_IMPORTS_PYTHON` names the interpreter whose standard library the sweep runs, `PROSE_SETTLE_WIDTHS` adds widths beside the default, `PROSE_IMPORTS_TIMEOUT` bounds one module's run, and `PROSE_IMPORTS_BAKE` writes the break set into the file it names. Every other run ratchets against the set tracked at `crate/tests/imports/baseline.json`, failing on a break that set does not carry or on a module that has fallen out of comparison since it was baked, and failing outright where nothing readable sits there.
 
 The fixture tree is itself a corpus, and the same probe sweeps it on every `cargo test`, failing the suite on any subset that needs a second pass at any swept `code-line-length`, because a subset that settles at one budget can still edit its own output at another. A pair the registry's independence table declares independent also runs as one pipeline splicing both rules into a single buffer, failing the suite wherever that run differs from the chained one, and a pointed sweep reports how many files each undeclared pair's spliced run agreed on, which is the evidence a table entry rests on. `rust:subsets` is that probe aimed at a wider corpus, narrowed to the shipped default budget so the pointed sweep's wall clock holds, with `PROSE_SETTLE_RULES` keeping only the subsets that touch the rules it names, `PROSE_SETTLE_PAIRS` choosing whether a scoped run claims every pair touching its rules or only the pairs its own rules open, `PROSE_SETTLE_SHARD` taking one `k/n` share of the pairs, and `PROSE_SETTLE_VERIFY` folding every pair beside its chained run and failing on a divergence. A set of runs whose scopes partition the catalog takes the second claim, so each pair is probed once across the set rather than once per endpoint. Adding the reported case to the tree therefore turns the report into a permanent guard in the same stroke.
 
-The `🪻 Corpus` workflow runs the sweeps against the pinned interpreter's standard library on every pull request touching the crate or the corpus binaries, the mutation pass as its own `🎨 Mutations` job beside `⚓ Settle`, restoring the stage `rust:bake` writes on each `main` push and the mutation set it writes once per revision of the generator, its manifest, or the lockfile, gating on the settle, mutation, import, and subset verdicts, rendering the delta into its job's log as a report, and running the subset probe as one row per rule family named by the family's badge, each scoped to the rules the pull request touches under `crate/src/rules/` where nothing else the workflow watches changed and carrying the whole family otherwise. The `🪟 Imports` job reads the break set tracked in the repository and fails only on a break that set does not carry, so a module `main` already breaks does not fail a pull request that leaves it alone.
+The `🪻 Corpus` workflow runs the sweeps against the pinned interpreter's standard library on every pull request touching the crate or the corpus binaries, the mutation pass as its own `🎨 Mutations` job beside `⚓ Settle`, restoring the stage `rust:bake` writes on each `main` push and the mutation set it writes once per revision of the generator, its manifest, or the lockfile, gating on the settle, mutation, import, and subset verdicts, rendering the delta into its job's log as a report, and running the subset probe as one row per rule family named by the family's badge, each scoped to the rules the pull request touches under `crate/src/rules/` where nothing else the workflow watches changed and carrying the whole family otherwise. The `🪟 Imports` job reads the break set tracked in the repository and fails only on a break that set does not carry or on a module that has fallen out of comparison since it was baked, so a module `main` already breaks does not fail a pull request that leaves it alone.
 
 ---
